Small cleanup and refactor
