+++ libs/portals/admin/ids-admin/src/context/EnvironmentContext.tsx
@@ -11,10 +11,6 @@ export const EnvironmentContext = createContext<
   EnvironmentContextValues | undefined
 >(undefined)
 
-export function createEnvironmentContext() {
-  return createContext<EnvironmentContextValues | undefined>(undefined)
-}
-
 export const EnvironmentProvider: FC<EnvironmentContextValues> = ({
   availableEnvironments,
   selectedEnvironment,
