+++ libs/portals/admin/ids-admin/src/screens/Client/components/ClientsUrl.tsx
@@ -5,7 +5,7 @@ import { useLocale } from '@island.is/localization'
 
 import { m } from '../../../lib/messages'
 import { useEnvironmentState } from '../../../hooks/useEnvironmentState'
-import { ClientFormTypes } from '../EditClient.action'
+import { ClientFormTypes } from '../EditClient.schema'
 import { useErrorFormatMessage } from '../../../hooks/useFormatErrorMessage'
 import { useClient } from '../ClientContext'
 import { FormCard } from '../../../components/FormCard'
