+++ libs/portals/admin/ids-admin/src/screens/Permission/components/PermissionContent.tsx
@@ -8,7 +8,7 @@ import { AuthAdminTranslatedValue } from '@island.is/api/schema'
 import { m } from '../../../lib/messages'
 import { FormCard } from '../../../components/FormCard'
 import { usePermission } from '../PermissionContext'
-import { PermissionFormTypes } from '../EditPermission.action'
+import { PermissionFormTypes } from '../EditPermission.schema'
 import { Languages } from '../../../utils/languages'
 import { useErrorFormatMessage } from '../../../hooks/useFormatErrorMessage'
 import { useEnvironmentState } from '../../../hooks/useEnvironmentState'
