+++ libs/portals/admin/ids-admin/src/types/dynamicEnvironmentResult.ts
@@ -1,0 +1,7 @@
+import { AuthAdminEnvironment } from '@island.is/api/schema'
+
+export type DynamicEnvironmentResult<T> = {
+  [K in keyof T]: T[K]
+} & {
+  environment: AuthAdminEnvironment
+}
